tests/submission/Google/video/autoplay/autoplay_reflects_false.html

<!doctype html>
<html>
 <head>
  <title>video.autoplay - reflection test</title>
  <script src="../../../../resources/testharness.js"></script>
  <script src="../../../../resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-autoplay">spec reference</a></p>
  <video id="video">
  </video>
  <div id="log"></div>
  <script>
test(function() {
 assert_false(
  document.getElementById("video").autoplay,
  "videoElement.autoplay reflects false when absent");
});
  </script>
 </body>
</html>
